Update README.rst
artnet2twinkly.com is my site and it does not support https.  This link needs changing because it's throwing an unsafe warning because the server i'm borrowing is registered as a different url in its certificate.  I also thought the fact you can now manually edit your layouts is relevant info.